pyproject.toml
[tool.poetry]
name = "midjourney-api"
version = "0.1.0"
description = ""
authors = [
    "https://github.com/yokonsan",
    "markshawn2020 <shawninjuly@gmail.com>"
]
readme = "README.md"
packages = [{ include = "midjourney_api" }]

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "2.2.3"
fastapi = "0.95"
python-dotenv = "1.0.0"
uvicorn = "0.22.0"
pydantic = ">=1.10.7,<1.11.0"
aiohttp = ">=3.8.4,<3.9.0"
loguru = "0.7.0"
aiofiles = "23.1.0"
fastapi-socketio = "^0.0.10"
discord = "^2.2.3"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
